Second commit

+++ app/uk/gov/hmrc/calculatepublicpensionadjustment/models/calculation/OutOfDatesTaxYearsCalculation.scala
@@ -25,7 +25,7 @@ case class OutOfDatesTaxYearsCalculation(
   chargePaidByMember: Double,
   chargePaidBySchemes: Double,
   revisedChargableAmountBeforeTaxRate: Int,
-  revisedChargableAmountAfterTaxRate: Int,
+  revisedChargableAmountAfterTaxRate: Double,
   unusedAnnualAllowance: Int,
   taxYearSchemes: List[OutOfDatesTaxYearSchemeCalculation],
   adjustedCompensation: Option[Double]
@@ -44,7 +44,7 @@ object OutOfDatesTaxYearsCalculation {
         (__ \ "chargePaidByMember").read[Double] and
         (__ \ "chargePaidBySchemes").read[Double] and
         (__ \ "revisedChargableAmountBeforeTaxRate").read[Int] and
-        (__ \ "revisedChargableAmountAfterTaxRate").read[Int] and
+        (__ \ "revisedChargableAmountAfterTaxRate").read[Double] and
         (__ \ "unusedAnnualAllowance").read[Int] and
         (__ \ "taxYearSchemes").read[List[OutOfDatesTaxYearSchemeCalculation]] and
         (__ \ "adjustedCompensation").readNullable[Double]
@@ -63,7 +63,7 @@ object OutOfDatesTaxYearsCalculation {
         (__ \ "chargePaidByMember").write[Double] and
         (__ \ "chargePaidBySchemes").write[Double] and
         (__ \ "revisedChargableAmountBeforeTaxRate").write[Int] and
-        (__ \ "revisedChargableAmountAfterTaxRate").write[Int] and
+        (__ \ "revisedChargableAmountAfterTaxRate").write[Double] and
         (__ \ "unusedAnnualAllowance").write[Int] and
         (__ \ "taxYearSchemes").write[List[OutOfDatesTaxYearSchemeCalculation]] and
         (__ \ "adjustedCompensation").writeNullable[Double]
